requirements.txt
flask==3.0.2
flask-cors==4.0.0
opencv-python==4.9.0.80
mediapipe==0.10.21
tensorflow==2.17.1
numpy==1.26.4
gunicorn
pillow